FEAT fetch data from external storage (#3)
The weights files are too large for github and PyPI.
git lfs is not an alternative, as github gives only very tiny quotas.
Therefore, let's storge the files on our own server,
and fetch when necessary. Data is cached in DEEPNOG_DATA dir,
which is ~/deepnog_data/ per default, or may be set as env variable.